refactor: extract webauthn signin and elevate logic into util functions
